@@ -1,59 +1,58 @@
 {
-  "name": "welldone-my-locations",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^16.11.26",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.12",
-    "lodash": "^4.17.21",
-    "materialize-css": "^1.0.0-rc.2",
-    "node-sass": "^7.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-materialize": "^3.10.0",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "redux": "^4.1.2",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.1",
-    "typescript": "^4.6.2",
-    "uuid": "^8.3.2",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/lodash": "^4.14.179",
-    "@types/materialize-css": "^1.0.11",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/uuid": "^8.3.4"
-  }
+	"name": "my-locations",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"@reduxjs/toolkit": "^2.0.1",
+		"@testing-library/jest-dom": "^5.17.0",
+		"@testing-library/react": "^13.4.0",
+		"@testing-library/user-event": "^13.5.0",
+		"@types/jest": "^27.5.2",
+		"@types/node": "^16.18.68",
+		"@types/react": "^18.2.45",
+		"@types/react-dom": "^18.2.18",
+		"lodash": "^4.17.21",
+		"materialize-css": "^1.0.0",
+		"node-sass": "^9.0.0",
+		"react": "^18.2.0",
+		"react-bootstrap": "^2.9.2",
+		"react-dom": "^18.2.0",
+		"react-materialize": "^3.10.0",
+		"react-redux": "^9.0.4",
+		"react-router-dom": "^6.21.1",
+		"react-scripts": "5.0.1",
+		"typescript": "^4.9.5",
+		"uuid": "^9.0.1",
+		"web-vitals": "^2.1.4"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject"
+	},
+	"eslintConfig": {
+		"extends": [
+			"react-app",
+			"react-app/jest"
+		]
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@types/lodash": "^4.14.202",
+		"@types/materialize-css": "^1.0.14",
+		"@types/react-router-dom": "^5.3.3",
+		"@types/uuid": "^9.0.7"
+	}
 }
