@@ -1,87 +1,91 @@
-import { useRef } from "react";
+import { useRef } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { Button, Icon } from 'react-materialize'
+import { RootState, useAppDispatch, useAppSelector } from '../../store/store'
+import { CategoryModal, CategoryTitleBar, List, Toolbar } from '../index'
 import {
-  CategoryModal,
-  CategoryTitleBar,
-  List,
-  NavBtn,
-  Toolbar,
-} from "../index";
-import { useDispatch, useSelector } from "react-redux";
-import {
-  clearCurrentCategory,
-  deleteCategory,
-  setCurrentCategory,
-} from "../../actions/categoryActions";
-import { RootState } from "../../reducers";
-import { ICategory } from "../../entityTypes/ICategory";
-import { useOutsideOfAreaClick } from "../../utils/useOutsideOfAreaClick";
+	clearCurrentCategory,
+	deleteCategory,
+	setCurrentCategory,
+} from '../../actions/categoryActions'
+import { ICategory } from '../../entityTypes/ICategory'
+import { useOutsideOfAreaClick } from '../../hooks/useOutsideOfAreaClick'
 
 const CategoriesPage = () => {
-  const wrapperRef = useRef<HTMLDivElement>(null);
-  useOutsideOfAreaClick(wrapperRef, clearCurrentCategory);
+	const navigate = useNavigate()
 
-  const categoriesState = useSelector(
-    (state: RootState) => state.categoriesReducer
-  );
-  const {
-    categories,
-    currentCategory,
-  }: { categories: ICategory[]; currentCategory: ICategory } = categoriesState;
-  const dispatch = useDispatch();
+	const wrapperRef = useRef<HTMLDivElement>(null)
+	useOutsideOfAreaClick(wrapperRef, clearCurrentCategory)
 
-  const deleteFunction = () => {
-    dispatch(deleteCategory(currentCategory.id));
-    dispatch(clearCurrentCategory());
-  };
+	const categoriesState = useAppSelector(
+		(state: RootState) => state.categoriesReducer
+	)
+	const { categories, currentCategory } = categoriesState
+	const dispatch = useAppDispatch()
 
-  const onItemClick = (selectedItem: ICategory) => {
-    dispatch(setCurrentCategory(selectedItem));
-  };
+	const deleteFunction = () => {
+		dispatch(deleteCategory(currentCategory.id))
+		dispatch(clearCurrentCategory())
+	}
 
-  return (
-    <div className="container" ref={wrapperRef}>
-      <div className="section">
-        <Toolbar
-          addTooltipMsg="Add a new category"
-          deleteFunction={deleteFunction}
-          deleteTooltipMsg="Delete category"
-          selectedItem={currentCategory}
-        >
-          {({
-            isEditMode,
-            isOpen,
-            isViewMode,
-            selectedItem,
-            setIsEditMode,
-            setIsOpen,
-            setIsViewMode,
-          }) => (
-            <>
-              <CategoryTitleBar
-                isEditMode={isEditMode}
-                setIsEditMode={setIsEditMode}
-              />
-              <CategoryModal
-                isOpen={isOpen}
-                isViewMode={isViewMode}
-                selectedCategory={selectedItem}
-                setIsOpen={setIsOpen}
-                setIsViewMode={setIsViewMode}
-              />
-            </>
-          )}
-        </Toolbar>
-        <List
-          emptyMsg="List is currently empty. Add some categories."
-          listItems={categories}
-          onItemClick={onItemClick}
-          selectedItem={currentCategory}
-        >
-          {(item) => <NavBtn currentCategory={item} />}
-        </List>
-      </div>
-    </div>
-  );
-};
+	const onItemClick = (selectedItem: ICategory) => {
+		dispatch(setCurrentCategory(selectedItem))
+	}
 
-export default CategoriesPage;
+	return (
+		<div className="container" ref={wrapperRef}>
+			<div className="section">
+				<Toolbar
+					addTooltipMsg="Add a new category"
+					deleteFunction={deleteFunction}
+					deleteTooltipMsg="Delete category"
+					selectedItem={currentCategory}
+				>
+					{({
+						isEditMode,
+						isOpen,
+						isViewMode,
+						selectedItem,
+						setIsEditMode,
+						setIsOpen,
+						setIsViewMode,
+					}) => (
+						<>
+							<CategoryTitleBar
+								isEditMode={isEditMode}
+								setIsEditMode={setIsEditMode}
+							/>
+							<CategoryModal
+								isOpen={isOpen}
+								isViewMode={isViewMode}
+								selectedCategory={selectedItem}
+								setIsOpen={setIsOpen}
+								setIsViewMode={setIsViewMode}
+							/>
+						</>
+					)}
+				</Toolbar>
+				<List
+					emptyMsg="List is currently empty. Add some categories."
+					listItems={categories}
+					onItemClick={onItemClick}
+					selectedItem={currentCategory}
+				>
+					{(item) => (
+						<Button
+							flat
+							icon={<Icon>send</Icon>}
+							node="button"
+							onClick={() => {
+								navigate(`/${item?.name}/locations`)
+							}}
+							tooltip={`View Locations for ${item?.name}`}
+						/>
+					)}
+				</List>
+			</div>
+		</div>
+	)
+}
+
+export default CategoriesPage
