@@ -1,16 +1,12 @@
-export { default as Navbar } from "./Navbar/Navbar";
-export { default as CategoriesPage } from "./CategoriesPage/CategoriesPage";
-export { default as LocationsPage } from "./LocationsPage/LocationsPage";
-export { default as AboutPage } from "./AboutPage/AboutPage";
-export { default as Toolbar } from "./Toolbar/Toolbar";
-export { default as CategoryTitleBar } from "./CategoryTitleBar/CategoryTitleBar";
-export { default as LocationTitleBar } from "./LocationTitleBar/LocationTitleBar";
-export { default as List } from "./List/List";
-export { default as ListItem } from "./ListItem/ListItem";
-export { default as AddBtn } from "./AddBtn/AddBtn";
-export { default as ViewDetailsBtn } from "./ViewDetailsBtn/ViewDetailsBtn";
-export { default as EditBtn } from "./EditBtn/EditBtn";
-export { default as DeleteBtn } from "./DeleteBtn/DeleteBtn";
-export { default as NavBtn } from "./NavBtn/NavBtn";
-export { default as CategoryModal } from "./CategoryModal/CategoryModal";
-export { default as LocationModal } from "./LocationModal/LocationModal";
+export { default as Navbar } from './Navbar/Navbar'
+export { default as CategoriesPage } from './CategoriesPage/CategoriesPage'
+export { default as LocationsPage } from './LocationsPage/LocationsPage'
+export { default as AboutPage } from './AboutPage/AboutPage'
+export { default as Toolbar } from './Toolbar/Toolbar'
+export { default as CategoryTitleBar } from './CategoryTitleBar/CategoryTitleBar'
+export { default as LocationTitleBar } from './LocationTitleBar/LocationTitleBar'
+export { default as List } from './List/List'
+export { default as ListItem } from './ListItem/ListItem'
+export { default as EditBtn } from './EditBtn/EditBtn'
+export { default as CategoryModal } from './CategoryModal/CategoryModal'
+export { default as LocationModal } from './LocationModal/LocationModal'
